# Fix loading AWS libraries to improve performance