Change initiate AWS SDK from handler to parser time. That can use native Node code-cache and this is not included to billed duration.

### Practical result:
Before
`Duration: 4931.01 ms	Billed Duration: 5000 ms Memory Size: 128 MB	Max Memory Used: 56 MB`
After
`Duration: 2601.95 ms	Billed Duration: 2700 ms Memory Size: 128 MB	Max Memory Used: 32 MB`